refactor: extract vaadin-overlay logic to the reusable mixin
